requirements.txt
blinker==1.3
chardet==2.3.0
cloud-init==18.2
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
dj-database-url==0.5.0
Django==1.11
djangorestframework==3.8.2
gunicorn==19.4.5
hibagent==1.0.1
idna==2.0
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
MarkupSafe==0.23
oauthlib==1.0.3
prettytable==0.7.2
psycopg2==2.7.4
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.1
python-debian==0.1.27
python-systemd==231
pytz==2018.4
PyYAML==3.11
requests==2.9.1
six==1.10.0
ssh-import-id==5.5
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtualenv==15.0.1
